add astype(Int64) to change data to integers clean_date_details

+++ src/data_cleaning.py
@@ -118,9 +118,9 @@ class DataCleaning:
 	def clean_date_details(self, date_details):
 		df = date_details
 		#day, month and year contains only digits
-		df["day"]=pd.to_numeric(df["day"], errors="coerce")
-		df["month"]=pd.to_numeric(df["month"], errors="coerce")
-		df["year"]=pd.to_numeric(df["year"], errors="coerce")
+		df["day"]=pd.to_numeric(df["day"], errors="coerce").astype('Int64')
+		df["month"]=pd.to_numeric(df["month"], errors="coerce").astype('Int64')
+		df["year"]=pd.to_numeric(df["year"], errors="coerce").astype(Int64)
 		#Time-period: remove anything that isn't one of ['Morning', 'Midday', 'Late_Hours', 'Evening']
 		df = df[df["time_period"].str.contains("Morning|Midday|Evening|Late_Hours")]
 		#uuid: keep only lines that do not contain capital letters
